added test for clicking next
